requirements.txt
bcrypt~=4.2.1
Flask~=3.1.0
pycryptodome~=3.14.1
pytest~=8.3.4
setuptools~=75.1.0
numpy~=1.26.4
openpyxl~=3.1.5
pillow~=10.4.0
python-docx~=1.1.2
pymongo~=4.10.1